+++ parachain/pallets/ethereum-beacon-client/tests/fixtures/execution-header-update.minimal.json
