+++ parachain/pallets/ethereum-beacon-client/tests/fixtures/finalized-header-update.minimal.json
