+++ parachain/pallets/ethereum-beacon-client/tests/fixtures/initial-checkpoint.mainnet.json
