--- parachain/pallets/ethereum-beacon-client/tests/fixtures/ssz_test_minimal_beacon_block_body.json
@@ -1,53 +1,0 @@
-{
-  "randao_reveal": "0x89dcaa27c0b62d269621e26e03860cf1f5da5866757502d6c4556739f42d6a6824bcb201fdf24b260b1196fa217935470e7182c7c53fadebddb8fe3379b9f5f855b9f2e30f98b574ae77457b3cae68a7d43c535aa98ee8bc34d3a9faa8ce0aee",
-  "eth1_data": {
-    "deposit_root": "0x6a0f9d6cb0868daa22c365563bb113b05f7568ef9ee65fdfeb49a319eaf708cf",
-    "deposit_count": 8,
-    "block_hash": "0x6f13d8e3bc84f30d3cd546587cde300b4e60890e5ee14c9529fd27b83cf09361"
-  },
-  "graffiti": "0x4c6f6465737461722d76312e342e322f636c6172612f534e4f2d3338392d6164",
-  "proposer_slashings": [],
-  "attester_slashings": [],
-  "attestations": [
-    {
-      "aggregation_bits": "0x03",
-      "data": {
-        "slot": 4317,
-        "index": 0,
-        "beacon_block_root": "0x6f054ea7ce689df194425798937714e1a35cfffe954b75efcad81db32972ae17",
-        "source": {
-          "epoch": 538,
-          "root": "0xd81b7781246ffd772f08294ae6d593399a903b059ec770dd176f00f45b3d7be2"
-        },
-        "target": {
-          "epoch": 539,
-          "root": "0x8241bc63edb3d425ce6d175513106f002d5f4f603ae7cb1cd25fffd24584f15f"
-        }
-      },
-      "signature": "0xb4d0e769388742a365bc59f820a336af9e87883285a659a5ec6321892e559ddada929f4d583bdb7efd9eca26523b39d7074be1ee9aa6da56506a89aad4e0c8d69f5c38184eac31885236bb517510132881579304fe6eee2b510656c195ef80bc"
-    }
-  ],
-  "deposits": [],
-  "voluntary_exits": [],
-  "sync_aggregate": {
-    "sync_committee_bits": "0xffffffff",
-    "sync_committee_signature": "0xb2aa795ca0140510eee21279d9bbc3695853c62906d23e36669e4d2d989846f70740e22c976dac8f4811be8b9e46bb5a062d6aebc46c884a4606b768263f12a8e1b7d6b1b6c129baafed3a508260d7bf2a0da51e83ee57e25e7e7b25d1856fd7"
-  },
-  "execution_payload": {
-    "parent_hash": "0x48516d7969e010750fcfeb96205abfc1409bc58d7be546135bfebddd01c18751",
-    "fee_recipient": "0x0000000000000000000000000000000000000000",
-    "state_root": "0x2c988c2df2d94e57532ff5917134efc06d701b13209ab1f3584d9134ff162577",
-    "receipts_root": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
-    "logs_bloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "prev_randao": "0x9e4f90822e09fc0089eeaad4727b04af7fe86a291ba86f82f6c8bf234a27c1f9",
-    "block_number": 4318,
-    "gas_limit": 30000000,
-    "gas_used": 0,
-    "timestamp": 1676898944,
-    "extra_data": "0x",
-    "base_fee_per_gas": 7,
-    "block_hash": "0x76e697cb4e9e9e221721cbe0e7ba3bbe5333a2751745101da1412c2b96ea9567",
-    "transactions_root": "0x7ffe241ea60187fdb0187bfa22de35d1f9bed7ab061d9401fd47e34a54fbede1"
-  }
-}
-
