+++ parachain/pallets/ethereum-beacon-client/tests/fixtures/sync-committee-update.mainnet.json
