+++ parachain/pallets/ethereum-beacon-client/tests/fixtures/sync-committee-update.minimal.json
