--- parachain/pallets/ethereum-beacon-client/tests/fixtures/szz_test_attester_slashing.json
@@ -1,155 +1,0 @@
-{
-    "attestation_1": {
-        "attesting_indices": [
-            1176,
-            1948,
-            2069,
-            2325,
-            3683,
-            4452,
-            4494,
-            6259,
-            7232,
-            7846,
-            8214,
-            8450,
-            9358,
-            9794,
-            14161,
-            15410,
-            15822,
-            16657,
-            17274,
-            17679,
-            18131,
-            19049,
-            19975,
-            20014,
-            20642,
-            21218,
-            22363,
-            22494,
-            22562,
-            22750,
-            23812,
-            23857,
-            25634,
-            25657,
-            26315,
-            26678,
-            27582,
-            27833,
-            40745,
-            41104,
-            41952,
-            42870,
-            43646,
-            43976,
-            44782,
-            45526,
-            46136,
-            46611,
-            46747,
-            46749,
-            46950,
-            47656,
-            48157,
-            48773,
-            50267,
-            52263,
-            54548,
-            55479,
-            55499,
-            55525,
-            56957,
-            57118,
-            58532,
-            58991,
-            59891,
-            60453,
-            60990,
-            61113,
-            62874,
-            62963,
-            63011,
-            63712,
-            64026,
-            64041,
-            66987,
-            67643,
-            70019,
-            70203,
-            74810,
-            75753,
-            75802,
-            76320,
-            76793,
-            77347,
-            77761,
-            78842,
-            79467,
-            82180,
-            82244,
-            82397,
-            83156,
-            83285,
-            83845,
-            84211,
-            84296,
-            84861,
-            85467,
-            87004,
-            87380,
-            87721,
-            87922,
-            88295,
-            88876,
-            90321,
-            90701,
-            91776,
-            92449,
-            92471,
-            93273,
-            93931,
-            94152,
-            94412,
-            95006,
-            99021,
-            100440,
-            100609
-        ],
-        "data": {
-            "slot": 45552,
-            "index": 16,
-            "beacon_block_root": "0x771e7e7eaae2bfc3e1e1ba73429f9d2c4a2bbdf1c0fc2617b1d4a16609ea7550", 
-            "source": {
-                "epoch": 1422,
-                "root": "0x541e8eea601b59ef610f86a0720ca56eb2ecec7de828d560ab70129d8e63bc34"
-            },
-            "target": {
-                "epoch": 1423,
-                "root": "0x6ab7cf2f6ccc42c64ff413619ffb3c5082ae29cb8e8a19c98c99599f21d4e5ca"
-            }
-        },
-        "signature": "0x83493fae00397611f6c185682c45797d471b8a85edae11430d93cf49a3b748c109ca0d41af02a108c8b85bcd909b6b220d07a6aa0ba200da071dd6176c7eca457b61f094d72ee425a0a56735c9e0bd63c77bf277049ff4f457cea7ed59a8bae4"
-    },
-    "attestation_2": {
-        "attesting_indices": [
-            100609
-        ],
-        "data": {
-            "slot": 45552,
-            "index": 16,
-            "beacon_block_root": "0xeed1acd77c359b5b02ca9ff13f76f816f3cd6616a3488aee74e63c25fc47ddfe", 
-            "source": {
-                "epoch": 1422,
-                "root": "0x541e8eea601b59ef610f86a0720ca56eb2ecec7de828d560ab70129d8e63bc34"
-            },
-            "target": {
-                "epoch": 1423,
-                "root": "0x6ab7cf2f6ccc42c64ff413619ffb3c5082ae29cb8e8a19c98c99599f21d4e5ca"
-            }
-        },
-        "signature": "0xafb529132db753a97e03102f5ff8c77c549b02885db0abe6899f744145aee90810855f81b6916bce1eafba4e25dd3f41071c38d998cdb3bcee219c9d89a301ffb2d1748a0ae354780f7ede0b46ccba809544b0d815281e586ddde2296d473df9"
-    }
-}
